+++ rpn.hs
@@ -155,7 +155,7 @@ comprimeEsto linea vars = Just $ Variable nombre comprimido : tail vars
 
 
 -- forzar a usar punto flotante al convertir a números
-read' :: String-> Float
+read' :: String-> Double
 read' = read
 
 -- Lo anterior obliga a definir división entera y módulos en puntos flotantes
